refactor(trunk-ir): move canonicalize patterns to per-dialect registries
Closes #690.

Each dialect now owns its canonicalization patterns next to its op
definitions. The `transforms::canonicalize` pass becomes a thin shell
that aggregates patterns from each dialect via free functions:

- `arith::canonicalization_patterns()` — `AddZeroFold`, `SubZeroFold`,
  `MulOneFold`, `MulZeroFold`, `IntConstFold` plus their
  arith-const-walking helpers (`const_int_def`, `core_int_width`,
  `wrap_signed_to_width`, ...).
- `core::canonicalization_patterns()` — `UnrealizedCastIdentity`,
  `UnrealizedCastRoundTrip`.

Pattern-specific tests move with each pattern into the dialect
modules (snapshot file paths shift accordingly). The pass's own tests
— region walk, fixpoint convergence, multi-pattern collaboration —
stay in `canonicalize.rs` since they exercise the pass shell rather
than any one pattern.

Adds `PatternApplicator::add_pattern_box` so the aggregator can
fold a `Vec<Box<dyn RewritePattern>>` into the applicator without
naming each concrete type at the call site.

Adding a new pattern to an existing dialect now touches exactly that
dialect's file. Adding a brand-new dialect's patterns is a one-line
`chain` addition in `canonicalize`.

No behavior change: same pattern set, same pass position in
`run_cleanup_passes`.
